--- a/dbms/src/Interpreters/sortBlock.cpp
+++ b/dbms/src/Interpreters/sortBlock.cpp
@@ -16,7 +16,13 @@
 #include <Columns/ColumnString.h>
 #include <Common/typeid_cast.h>
 #include <Interpreters/sortBlock.h>
+#include <Storages/Transaction/Collator.h>
+#include <Storages/Transaction/CollatorUtils.h>
+#include <common/defines.h>
 
+#if defined(APPLY_FOR_TYPE) || defined(M) || defined(CONCAT)
+static_assert(false);
+#endif
 
 namespace DB
 {
@@ -46,8 +52,7 @@ static ColumnsWithSortDescriptions getColumnsWithSortDescription(const Block & b
     return res;
 }
 
-
-static inline bool needCollation(const IColumn * column, const SortColumnDescription & description)
+ALWAYS_INLINE static inline bool NeedCollation(const IColumn * column, const SortColumnDescription & description)
 {
     if (!description.collator)
         return false;
@@ -62,6 +67,229 @@ static inline bool needCollation(const IColumn * column, const SortColumnDescrip
     return true;
 }
 
+#define APPLY_FOR_TYPE(M) \
+    M(UInt64)             \
+    M(Int64)              \
+    M(StringBinPadding)   \
+    M(StringBinNoPadding) \
+    M(StringWithCollatorGenetic)
+
+#define CONCAT(x, y) x##y
+
+#define M(NAME) NAME,
+enum class FastPathType
+{
+    APPLY_FOR_TYPE(M)
+};
+#undef M
+
+constexpr size_t max_fast_path_num = 2;
+
+template <typename T>
+struct ColumnVecCompare
+{
+    static inline const ColumnVector<T> * intoTarget(const IColumn * column)
+    {
+        return static_cast<const ColumnVector<T> *>(column);
+    }
+
+    template <size_t index>
+    ALWAYS_INLINE static inline int compareAt(size_t a, size_t b, const ColumnsWithSortDescriptions & columns_with_sort_desc)
+    {
+        const auto & desc = columns_with_sort_desc[index];
+        const auto * column = intoTarget(desc.first);
+        return column->compareAt(a, b, *column, 0) * desc.second.direction;
+    }
+};
+
+template <FastPathType type>
+struct ColumnStringCompare
+{
+    static inline const ColumnString * intoTarget(const IColumn * column)
+    {
+        return static_cast<const ColumnString *>(column);
+    }
+
+    template <size_t index>
+    ALWAYS_INLINE static inline int compareAt(size_t a, size_t b, const ColumnsWithSortDescriptions & columns_with_sort_desc)
+    {
+        const auto & desc = columns_with_sort_desc[index];
+        const auto * column = intoTarget(desc.first);
+        int ret = 0;
+        {
+            auto str_a = column->getDataAt(a);
+            auto str_b = column->getDataAt(b);
+
+            if constexpr (type == FastPathType::StringBinPadding)
+            {
+                ret = BinCollatorCompare<true>(str_a.data, str_a.size, str_b.data, str_b.size);
+            }
+            else if constexpr (type == FastPathType::StringBinNoPadding)
+            {
+                ret = BinCollatorCompare<false>(str_a.data, str_a.size, str_b.data, str_b.size);
+            }
+            else
+            {
+                ret = desc.second.collator->compare(str_a.data, str_a.size, str_b.data, str_b.size);
+            }
+        }
+        return desc.second.direction * ret;
+    }
+};
+
+using ColumnCompareUInt64 = ColumnVecCompare<UInt64>;
+using ColumnCompareInt64 = ColumnVecCompare<Int64>;
+using ColumnCompareStringBinPadding = ColumnStringCompare<FastPathType::StringBinPadding>;
+using ColumnCompareStringBinNoPadding = ColumnStringCompare<FastPathType::StringBinNoPadding>;
+using ColumnCompareStringWithCollatorGenetic = ColumnStringCompare<FastPathType::StringWithCollatorGenetic>;
+
+// only for uint64, int64, string
+template <typename ColumnCmpA, typename ColumnCmpB>
+struct MultiColumnSortFastPath
+{
+    const ColumnsWithSortDescriptions & columns_with_sort_desc;
+
+    explicit MultiColumnSortFastPath(const ColumnsWithSortDescriptions & columns_with_sort_desc_)
+        : columns_with_sort_desc(columns_with_sort_desc_)
+    {}
+
+    ALWAYS_INLINE inline int compareAt(size_t a, size_t b) const
+    {
+        constexpr size_t index_a = 0, index_b = 1;
+
+        int ret = 0;
+        {
+            ret = ColumnCmpA::template compareAt<index_a>(a, b, columns_with_sort_desc);
+        }
+        if (!ret)
+        {
+            ret = ColumnCmpB::template compareAt<index_b>(a, b, columns_with_sort_desc);
+        }
+        return ret;
+    }
+
+    ALWAYS_INLINE inline bool operator()(size_t a, size_t b) const
+    {
+        int ret = compareAt(a, b);
+        return ret < 0;
+    }
+};
+
+struct CollatorDesc : boost::noncopyable
+{
+    const ColumnsWithSortDescriptions & columns_with_sort_desc;
+    std::vector<bool> need_collations;
+
+    bool has_collation{false};
+
+    bool can_use_fast_path = false;
+    std::array<FastPathType, max_fast_path_num> type_for_fast_path{};
+    size_t fast_path_cnt = 0;
+
+    ALWAYS_INLINE static bool isUInt64(const IColumn * column)
+    {
+        return typeid_cast<const ColumnUInt64 *>(column);
+    }
+    ALWAYS_INLINE static bool isInt64(const IColumn * column)
+    {
+        return typeid_cast<const ColumnInt64 *>(column);
+    }
+    ALWAYS_INLINE static bool isString(const IColumn * column)
+    {
+        return typeid_cast<const ColumnString *>(column);
+    }
+
+    inline void addFastPathType(FastPathType tp)
+    {
+        type_for_fast_path[fast_path_cnt++] = tp;
+    }
+
+    explicit CollatorDesc(const ColumnsWithSortDescriptions & columns_with_sort_desc_)
+        : columns_with_sort_desc(columns_with_sort_desc_)
+    {
+        need_collations.reserve(columns_with_sort_desc.size());
+
+        can_use_fast_path = columns_with_sort_desc.size() == max_fast_path_num;
+
+        for (const auto & sort_desc : columns_with_sort_desc)
+        {
+            auto need_collation = DB::NeedCollation(sort_desc.first, sort_desc.second);
+            if (need_collation)
+            {
+                has_collation = true;
+                if (isString(sort_desc.first))
+                {
+                    // only when column is string(not nullable)
+                    auto collator_type = TiDB::GetTiDBCollatorType(sort_desc.second.collator);
+                    if (can_use_fast_path)
+                    {
+                        switch (collator_type)
+                        {
+                        case TiDB::ITiDBCollator::CollatorType::UTF8MB4_BIN:
+                        case TiDB::ITiDBCollator::CollatorType::UTF8_BIN:
+                        case TiDB::ITiDBCollator::CollatorType::LATIN1_BIN:
+                        case TiDB::ITiDBCollator::CollatorType::ASCII_BIN:
+                        {
+                            addFastPathType(FastPathType::StringBinPadding);
+                            break;
+                        }
+                        case TiDB::ITiDBCollator::CollatorType::BINARY:
+                        {
+                            addFastPathType(FastPathType::StringBinNoPadding);
+                            break;
+                        }
+                        default:
+                        {
+                            addFastPathType(FastPathType::StringWithCollatorGenetic);
+                            break;
+                        }
+                        }
+                    }
+                }
+                else
+                {
+                    // nullable
+                    can_use_fast_path = false;
+                }
+            }
+            else if (can_use_fast_path)
+            {
+                if (isUInt64(sort_desc.first))
+                {
+                    addFastPathType(FastPathType::UInt64);
+                }
+                else if (isInt64(sort_desc.first))
+                {
+                    addFastPathType(FastPathType::Int64);
+                }
+                else
+                {
+                    can_use_fast_path = false;
+                }
+            }
+            need_collations.emplace_back(need_collation);
+        }
+    }
+
+    ALWAYS_INLINE inline size_t size() const
+    {
+        return columns_with_sort_desc.size();
+    }
+
+    ALWAYS_INLINE inline int compareAt(size_t col_index, size_t a, size_t b) const
+    {
+        const auto & column = columns_with_sort_desc[col_index];
+
+        if (likely(!need_collations[col_index]))
+        {
+            return column.second.direction * column.first->compareAt(a, b, *column.first, column.second.nulls_direction);
+        }
+        else
+        {
+            return column.second.direction * column.first->compareAt(a, b, *column.first, column.second.nulls_direction, *column.second.collator);
+        }
+    }
+};
 
 struct PartialSortingLess
 {
@@ -71,48 +299,88 @@ struct PartialSortingLess
         : columns(columns_)
     {}
 
-    bool operator()(size_t a, size_t b) const
+    ALWAYS_INLINE inline bool operator()(size_t a, size_t b) const
     {
+        int res = 0;
         for (const auto & column : columns)
         {
-            int res = column.second.direction * column.first->compareAt(a, b, *column.first, column.second.nulls_direction);
-            if (res < 0)
-                return true;
-            else if (res > 0)
-                return false;
+            res = column.second.direction * column.first->compareAt(a, b, *column.first, column.second.nulls_direction);
+            if (res)
+                break;
         }
-        return false;
+        return res < 0;
     }
 };
 
 struct PartialSortingLessWithCollation
 {
-    const ColumnsWithSortDescriptions & columns;
+    const CollatorDesc & collator_desc;
 
-    explicit PartialSortingLessWithCollation(const ColumnsWithSortDescriptions & columns_)
-        : columns(columns_)
-    {}
-
-    bool operator()(size_t a, size_t b) const
+    explicit PartialSortingLessWithCollation(const CollatorDesc & collator_desc_)
+        : collator_desc(collator_desc_)
     {
-        for (const auto & column : columns)
-        {
-            int res;
-            if (needCollation(column.first, column.second))
-                res = column.first->compareAt(a, b, *column.first, column.second.nulls_direction, *column.second.collator);
-            else
-                res = column.first->compareAt(a, b, *column.first, column.second.nulls_direction);
+        assert(collator_desc.has_collation);
+    }
 
-            res *= column.second.direction;
-            if (res < 0)
-                return true;
-            else if (res > 0)
-                return false;
+    ALWAYS_INLINE inline bool operator()(size_t a, size_t b) const
+    {
+        int res = 0;
+        for (size_t i = 0; i < collator_desc.size(); ++i)
+        {
+            res = collator_desc.compareAt(i, a, b);
+            if (res)
+                break;
         }
-        return false;
+        return res < 0;
     }
 };
 
+template <typename F>
+ALWAYS_INLINE static inline void PermutationSort(IColumn::Permutation & perm, size_t limit, F && fn_cmp)
+{
+    if (limit)
+        std::partial_sort(perm.begin(), perm.begin() + limit, perm.end(), fn_cmp);
+    else
+        std::sort(perm.begin(), perm.end(), fn_cmp);
+}
+
+template <typename A, typename B>
+void FastPathPermutationSort_P2(const CollatorDesc & desc, IColumn::Permutation & perm, size_t limit)
+{
+    MultiColumnSortFastPath<A, B> cmp{desc.columns_with_sort_desc};
+    return PermutationSort(perm, limit, cmp);
+}
+
+template <typename A>
+ALWAYS_INLINE inline void FastPathPermutationSort_P1(const CollatorDesc & desc, IColumn::Permutation & perm, size_t limit)
+{
+#define M(NAME)                                                                               \
+    case FastPathType::NAME:                                                                  \
+    {                                                                                         \
+        return FastPathPermutationSort_P2<A, CONCAT(ColumnCompare, NAME)>(desc, perm, limit); \
+    }
+
+    switch (desc.type_for_fast_path[1])
+    {
+        APPLY_FOR_TYPE(M)
+    }
+#undef M
+}
+
+ALWAYS_INLINE inline void FastPathPermutationSort(const CollatorDesc & desc, IColumn::Permutation & perm, size_t limit)
+{
+#define M(NAME)                                                                            \
+    case FastPathType::NAME:                                                               \
+    {                                                                                      \
+        return FastPathPermutationSort_P1<CONCAT(ColumnCompare, NAME)>(desc, perm, limit); \
+    }
+
+    switch (desc.type_for_fast_path[0])
+    {
+        APPLY_FOR_TYPE(M)
+    }
+#undef M
+} // namespace DB
 
 void sortBlock(Block & block, const SortDescription & description, size_t limit)
 {
@@ -122,14 +390,14 @@ void sortBlock(Block & block, const SortDescription & description, size_t limit)
     /// If only one column to sort by
     if (description.size() == 1)
     {
-        bool reverse = description[0].direction == -1;
+        bool reverse = description[0].direction < 0;
 
         const IColumn * column = !description[0].column_name.empty()
             ? block.getByName(description[0].column_name).column.get()
             : block.safeGetByPosition(description[0].column_number).column.get();
 
         IColumn::Permutation perm;
-        if (needCollation(column, description[0]))
+        if (NeedCollation(column, description[0]))
             column->getPermutation(*description[0].collator, reverse, limit, description[0].nulls_direction, perm);
         else
             column->getPermutation(reverse, limit, description[0].nulls_direction, perm);
@@ -148,39 +416,26 @@ void sortBlock(Block & block, const SortDescription & description, size_t limit)
         if (limit >= size)
             limit = 0;
 
-        bool need_collation = false;
         ColumnsWithSortDescriptions columns_with_sort_desc = getColumnsWithSortDescription(block, description);
-
-        for (size_t i = 0, size = description.size(); i < size; ++i)
+        const auto collator_desc = CollatorDesc{columns_with_sort_desc};
+        if (collator_desc.can_use_fast_path)
         {
-            if (needCollation(columns_with_sort_desc[i].first, description[i]))
-            {
-                need_collation = true;
-                break;
-            }
+            assert(collator_desc.fast_path_cnt == max_fast_path_num);
+
+            FastPathPermutationSort(collator_desc, perm, limit);
         }
-
-        if (need_collation)
+        else if (collator_desc.has_collation)
         {
-            PartialSortingLessWithCollation less_with_collation(columns_with_sort_desc);
-
-            if (limit)
-                std::partial_sort(perm.begin(), perm.begin() + limit, perm.end(), less_with_collation);
-            else
-                std::sort(perm.begin(), perm.end(), less_with_collation);
+            PartialSortingLessWithCollation less_with_collation(collator_desc);
+            PermutationSort(perm, limit, less_with_collation);
         }
         else
         {
             PartialSortingLess less(columns_with_sort_desc);
-
-            if (limit)
-                std::partial_sort(perm.begin(), perm.begin() + limit, perm.end(), less);
-            else
-                std::sort(perm.begin(), perm.end(), less);
+            PermutationSort(perm, limit, less);
         }
 
-        size_t columns = block.columns();
-        for (size_t i = 0; i < columns; ++i)
+        for (size_t i = 0; i < block.columns(); ++i)
             block.safeGetByPosition(i).column = block.safeGetByPosition(i).column->permute(perm, limit);
     }
 }