### What problem does this PR solve?

Issue Number: ref #5294

### What is changed and how it works?

- optimize when number of sort key is 2 and type of sort key is one of 
  - UInt64
  - Int64
  - StringBinPadding
  - StringBinNoPadding
  - StringWithCollatorGeneric
- we can optimize the case that number of sort key is 3 later if necessary (there will be a lot of template instantiations)

### Benchmark

ENV
- tpch-100
- 1 tiflash
- limit cpu up to 2000%
- original commit: 6d0cbc8cc78b4048fb88e0845d6c9c2881e01515

---

SQL

```
EXPLAIN analyze SELECT ROW_NUMBER() OVER w1 FROM PART window w1 AS (PARTITION BY P_MFGR order by P_SIZE);
```

- sort key: `STR(utf8 collator),UINT64`

Time(s) | Original | Optimized |   |   | Improvement
-- | -- | -- | -- | -- | --
  | 8.16 | 7.18 |   |   |  
  | 8.22 | 6.94 |   |   |  
  | 8.29 | 7.35 |   |   |  
  | 8.34 | 7.04 |   |   |  
  | 8.24 | 7.37 |   |   | AVG(Original) / AVG(Optimized) - 1.0
AVG | 8.25 | 7.176 |   | Optimized : Original | 14.97%

----

SQL

```
EXPLAIN analyze SELECT ROW_NUMBER() OVER w1 FROM PART window w1 AS (PARTITION BY P_MFGR ORDER BY p_name);
```

- sort key: `STR(utf8 collator),STR(utf8 collator)`

Time(s) | Original | Optimized |   |   | Improvement
-- | -- | -- | -- | -- | --
  | 18.3 | 16.34 |   |   |  
  | 18.27 | 16.62 |   |   |  
  | 18.7 | 16.57 |   |   |  
  | 18.57 | 16.52 |   |   |  
  | 18.55 | 16.63 |   |   | AVG(Original) / AVG(Optimized) - 1.0
AVG | 18.478 | 16.536 |   | Optimized : Original | 11.74%

----

SQL

```
EXPLAIN analyze SELECT ROW_NUMBER() OVER w1 FROM PART window w1 AS (PARTITION BY p_name ORDER BY p_partkey);
```

- sort key: `INT64,STR(utf8 collator)`
- costs of block sort are about 30%

Time(s) | Original | Optimized |   |   | Improvement
-- | -- | -- | -- | -- | --
  | 9.71 | 9.28 |   |   |  
  | 9.62 | 9.41 |   |   |  
  | 9.76 | 9.25 |   |   |  
  | 9.73 | 9.26 |   |   |  
  | 9.67 | 9.28 |   |   | AVG(Original) / AVG(Optimized) - 1.0
AVG | 9.698 | 9.296 |   | Optimized : Original | 4.32%


### Check List

Tests <!-- At least one of them must be included. -->

- [ ] Unit test
- [ ] Integration test
- [ ] Manual test (add detailed scripts or steps below)
- [ ] No code

Side effects

- [ ] Performance regression: Consumes more CPU
- [ ] Performance regression: Consumes more Memory
- [ ] Breaking backward compatibility

Documentation

- [ ] Affects user behaviors
- [ ] Contains syntax changes
- [ ] Contains variable changes
- [ ] Contains experimental features
- [ ] Changes MySQL compatibility

### Release note

<!-- bugfix or new feature needs a release note -->

```release-note
None
```