@solotzg 2022-09-21
> It seems that the optimization is all from move branches out of loops and make branches static?
> 
> from the offline discussion, we can materialize the collator look-up column, to decrease the collator look-up times from O(nlogn) to O(n)

It may also be an optional way if the overhead about materialization is small.