@solotzg 2022-09-22
> consider add a FastPathPermutationSort<1>, which is commoner than 2 columns

We already have such optimization :  https://github.com/pingcap/tiflash/blob/3674439a1f1aeb09c70ad67f70cf5c2e961bd01a/dbms/src/Interpreters/sortBlock.cpp#L123-L141