@solotzg 2022-09-23
/rebuild